+++ src/codem-ffmpeg.js
@@ -1,5 +1,6 @@
 const EventEmitter = require('events')
 const spawn        = require('child_process').spawn
+const fs           = require('fs')
 
 const ERRORS = {
   ArgumentError: new Error("Expected arguments to be an array"),
@@ -10,12 +11,29 @@ class FFmpeg extends EventEmitter {
   get ffmpegBinary ()          { return this._ffmpegBinary || 'ffmpeg' }
   set ffmpegBinary (newBinary) { this._ffmpegBinary = newBinary }
   
+  get attributes () {
+    return new Map([
+      ['filesize', this._filesize || null],
+      ['duration', this._duration],
+      ['inputFile', this._inputFile],
+      ['outputFile', this._outputFile]
+    ])
+  }
+
   get output () {
     return this._output || null
   }
-  get progress ()              {
+  get progress () {
     if (typeof this._duration === 'undefined' || typeof this._current === 'undefined') return null
     return Math.min(this._current / this._duration)
+  }
+
+  get _inputFile () {
+    return this._findFirstMatchInOutput(new RegExp(/^Input #\d+,.*, from '(.*)':$/))
+  }
+
+  get _outputFile () {
+    return this._findFirstMatchInOutput(new RegExp(/^Output #\d+,.*, to '(.*)':$/))
   }
   
   constructor(args) {
@@ -26,20 +44,20 @@ class FFmpeg extends EventEmitter {
   }
   
   cancel() {
-    if (this._child_process) {
-      this._child_process.kill('SIGINT');
+    if (this._childProcess) {
+      this._childProcess.kill('SIGINT');
     }
   }
   
   spawn() {
-    this._child_process = spawn(this.ffmpegBinary, this._args)
+    this._childProcess = spawn(this.ffmpegBinary, this._args)
 
-    this._child_process.stderr.on('data', (data)  => { this._processData(data) })
-    this._child_process.on('error', (err)         => { this.emit('error', ERRORS.SpawnError) })
-    this._child_process.on('exit', (code, signal) => { this.emit('exit', code, signal) })
+    this._childProcess.stderr.on('data', (data)  => { this._processData(data) })
+    this._childProcess.on('error', (err)         => { this.emit('error', ERRORS.SpawnError) })
+    this._childProcess.on('exit', (code, signal) => { this.emit('exit', code, signal) })
     
     this._output = ""
-    return this._child_process
+    return this._childProcess
   }
   
   _extractDuration(text) {
@@ -51,11 +69,27 @@ class FFmpeg extends EventEmitter {
 
     if (durationMatch == null) return
 
-    let [hours, minutes, seconds] = durationMatch.slice(1,4).map(t => parseInt(t, 10))
+    let [,hours, minutes, seconds] = durationMatch.map(t => parseInt(t, 10))
       
     this._duration = hours * 3600 + minutes * 60 + seconds
     this._current = 0
     this.emit('progress', this.progress)
+  }
+  
+  _extractFilesize() {
+    if (this._isStatRunning) return
+      
+    this._isStatRunning = true
+      
+    fs.stat(this._inputFile, (err, stats) => {
+      if (err) {
+        this._filesize = Number.NaN
+      } else {
+        this._filesize = (stats.isFile() ? stats.size : Number.NaN)
+      }
+      
+      this.isStatRunning = false
+    })
   }
   
   _extractProgress(text) {
@@ -64,13 +98,29 @@ class FFmpeg extends EventEmitter {
 
     if (timeMatch == null) return
     
-    let [hours, minutes, seconds] = timeMatch.slice(1,4).map(t => parseInt(t, 10))
+    let [,hours, minutes, seconds] = timeMatch.map(t => parseInt(t, 10))
     let newCurrent = hours * 3600 + minutes * 60 + seconds
 
     if (newCurrent == this._current) return
 
     this._current = newCurrent
     this.emit('progress', this.progress)
+  }
+  
+  _findFirstMatchInOutput(regexp) {
+    if (this.output === null) return null
+  
+    const lines = this.output.split("\n")
+    const inputMatcher = regexp
+
+    for (let line of lines) {
+      let inputMatch = inputMatcher.exec(line)
+      if (inputMatch && inputMatch[1]) {
+        return inputMatch[1]
+      }
+    }
+
+    return null
   }
   
   _processData(data) {
@@ -80,6 +130,10 @@ class FFmpeg extends EventEmitter {
       this._extractDuration(text)
     } else {
       this._extractProgress(text)
+    }
+    
+    if (typeof this._filesize === 'undefined' && this._inputFile != null) {
+      this._extractFilesize()
     }
   }
 }