+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## codem-ffmpeg 0.1.3 (2016/10/31) ##
+
+* Expose a custom `attributes` Map which contains educated guesses about the process
+
 ## codem-ffmpeg 0.1.2 (2016/10/21) ##
 
 * Add output property